Restructure jobs / handlers
