+++ src/jobs/jobs.js
@@ -1,15 +1,36 @@
-import { createLogger } from '../logging/logger.js'
-import { notifyStatusJob } from './notify-jobs.js'
+import semaphore from 'semaphore'
+import { CronJob } from 'cron'
 
-const logger = createLogger()
+import { config } from '../config/index.js'
+import { checkNotifyStatusHandler } from './check-notify-status/handler.js'
+
+const statusCheckMutex = semaphore(1)
+
+const notifyStatusJob = new CronJob(
+  config.get('jobs.checkNotifyStatus.cronPattern'),
+  async () => {
+    if (!statusCheckMutex.available(1)) {
+      console.log('Check notify status job already running')
+      return
+    }
+
+    statusCheckMutex.take(async () => {
+      try {
+        await checkNotifyStatusHandler()
+      } catch (error) {
+        console.error('Error running check notify status job:', error.message)
+      } finally {
+        statusCheckMutex.leave()
+      }
+    })
+  }
+)
 
 const startJobs = () => {
-  logger.info('Starting notification status check job')
   notifyStatusJob.start()
 }
 
 const stopJobs = () => {
-  logger.info('Stopping notification status check job')
   notifyStatusJob.stop()
 }
